+++ application/entity/sp_email_retry.php
@@ -3,7 +3,7 @@ defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
 'id' => 'int',
-'uid' => 'mediumint',
+'uid' => 'int',
 'email' => 'varchar',
 'tplcode' => 'varchar'
 );
