+++ application/entity/sp_member_ext.php
@@ -2,7 +2,7 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
-'uid' => 'mediumint',
+'uid' => 'int',
 'weixin' => 'varchar',
 'avatar' => 'varchar'
 );
