+++ application/entity/sp_member_login.php
@@ -2,7 +2,7 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
-'id' => 'mediumint',
+'id' => 'int',
 'uid' => 'int',
 'login_time' => 'int',
 'login_ip' => 'varchar',
