+++ application/entity/sp_site_email.php
@@ -2,7 +2,7 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
-'id' => 'mediumint',
+'id' => 'int',
 'email' => 'varchar',
 'title' => 'varchar',
 'content' => 'text',
