+++ application/entity/sp_verifycode_log.php
@@ -2,7 +2,7 @@
 defined('BASEPATH') OR exit('No direct script access allowed');
 
 $entity = array(
-'id' => 'mediumint',
+'id' => 'int',
 'phone' => 'varchar',
 'ip' => 'varchar',
 'code' => 'varchar',
